feat: constrain requirements-dev.in
Using the -r option previously only appended the packages listed in the
base requirements.in file to the requirements-dev.in. However, this
leads to incompatibilities between the two set of requirements as when
compiling the requirements-dev we get the most up to date packages
available instead of the constrained ones.