fix: do not remove contract from coverage report if a function is missing statements or branches
The previous behavior expects all functions to have a report for both
statements and branches. If not, it removes the whole contract from the
coverage report.
While the assumption seems reasonable, some functions seem to only have
either one of statements or branches, ending up in removing contracts
that would otherwise contain useful coverage information.

This is related to #1087 and should at least be a partial fix to it